changes to code, commenting and reorganization
1. noticed mistake in parsing the planet and mission
    (off by one) - should be fixed now.
2. switched open() and close() to the with idiom, but
    I need to double check that I'm doing it correctly :P
3. noticed inconsistency in the establishment of the last_id
    that is potentially creating string _id's for a few
    documents in the database. I think I have made them
    consistently integers now, but I need to check that
    it all works
3. lots of commenting and reorganization

unfortunately, because I do not have any test tweets saved
and a debug mode, I can't check that the parsing/database
populating code works unless I also have a connection. And
since I'm on a plane right now and I'm a poor college student
who is not going to pay for gogo inflight wifi, that's not
happening right now.

SO MAKE SURE YOU TEST AND THEN RECOMMIT WITH CHANGES